Parse dbpedia information.
Threaded approach to parse 500k fields from dbpedia location index.
Currently we are only capturing the name of the location, and the
lat/lon of the location itself.

Process benchmarked at 1778 updates / second.
Not bad, little script, not bad.

Next: parse out richer information (tags) to add to the schema.